server/requirements/common.txt
flask
flask-sqlalchemy
flask-script
redis
hiredis
flask-cache
flask-gzip

# for delayed jobs
#rq
